+++ R/summarizeExperiments.R
@@ -17,6 +17,6 @@ summarizeExperiments = function(ids = NULL, by = c("problem", "algorithm"), reg 
   assertCharacter(by, any.missing = FALSE, min.chars = 1L, min.len = 1L, unique = TRUE)
   tab = getJobPars(ids = ids, reg = reg)
   if (!setequal(by, c("problem", "algorithm")))
-    tab = flatten(tab)
+    tab = unwrap(tab)
   tab[, list(.count = .N), by = by]
 }